+++ RenX.Core/RenX_PlayerInfo.h
@@ -53,7 +53,7 @@ namespace RenX
 		Jupiter::StringS vehicle;
 		uint64_t steamid = 0;
 		uint32_t ip32 = 0;
-		uint8_t ban_flags = 0;
+		uint16_t ban_flags = 0;
 		TeamType team = TeamType::Other;
 		int id = 0;
 		bool isBot = false;
